Make servlet-api dependency 'provided'.
This is because there are multiple servlet-api maven artifacts out,
"javax.servlet:servlet-api:2.x" and "javax.servlet:javax.servlet-api:3.x"
and using the 3.x version would require users of this jar
to explicitly exclude the 2.x servlet api.
In this case it is easier for users of this jar to assume that
a valid servlet API is available at runtime.