Opentracing support dropped.
The opentracing context manager and timer implementations will not
be continued. The last-released version of these modules will be `2.0.2`.

These versions should remain compatible with all future `v2` api versions but will not be released in future versions of this library.

You can use the `contex-propagation-bom` (Bill of Materials), to keep track of latest releases of the various modules in this library.

Signed-off-by: Sjoerd Talsma <sjoerdtalsma@users.noreply.github.com>